Rename speedlimiter
